Address clippy complaints and move LayerLogger
New clippy complaints, specifically complaining about panics. I have some interfaces that are actually safe to use, but possibly unsafe within. I made them safer by hiding that interior state more. I moved that code in LayerLogger to its own module as it's already quite complex.

I moved ReadyYourWrite to a general module and I want to move all things that aren't BuildLog out of log.rs. Also revisit naming later.

Some clippy violations still exist due to placeholder code for the errors and one due to me knowing I'm going to revisit/re-write code on Monday with Manuel:

```
warning: usage of an `Arc` that is not `Send` or `Sync`
   --> commons/src/output/layer_logger.rs:112:20
    |
112 |             inner: Arc::new(Mutex::new(SectionLoggerCell::new(logger))),
    |                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    |
    = note: the trait `Send` is not implemented for `Mutex<SectionLoggerCell>`
    = note: the trait `Sync` is not implemented for `Mutex<SectionLoggerCell>`
    = note: required for `Arc<Mutex<SectionLoggerCell>>` to implement `Send` and `Sync`
    = help: consider using an `Rc` instead or wrapping the inner type with a `Mutex`
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#arc_with_non_send_sync
    = note: `#[warn(clippy::arc_with_non_send_sync)]` on by default
```
