+++ buildpacks/ruby/src/layers/bundle_download_layer.rs
@@ -2,8 +2,7 @@ use crate::RubyBuildpack;
 use crate::RubyBuildpackError;
 use commons::fun_run::{self, CommandWithName};
 use commons::gemfile_lock::ResolvedBundlerVersion;
-use commons::output::fmt;
-use commons::output::log::LayerLogger;
+use commons::output::{fmt, layer_logger::LayerLogger};
 use libcnb::build::BuildContext;
 use libcnb::data::layer_content_metadata::LayerTypes;
 use libcnb::layer::{ExistingLayerStrategy, Layer, LayerData, LayerResult, LayerResultBuilder};
