+++ buildpacks/ruby/src/rake_task_detect.rs
@@ -1,5 +1,5 @@
 use commons::fun_run::{self, CmdError, CommandWithName};
-use commons::output::{fmt, log::LayerLogger};
+use commons::output::{fmt, layer_logger::LayerLogger};
 use core::str::FromStr;
 use regex::Regex;
 use std::{ffi::OsStr, process::Command};
